+++ CHANGELOG.md
@@ -22,4 +22,5 @@ Formato de CHANGELOG baseado em [Keep a Changelog](https://keepachangelog.com/pt
 - Componente `Step Bar`
 - Componente `Table`
 - Componente `Tag`
+- Componente `Popover`
 - Componente `Radio`